+++ examples/test/dynarray/test_dynarray_literal_badelem.pas
@@ -1,0 +1,6 @@
+program TestDynArrayLiteralBadElem;
+var
+    arr: array of integer;
+begin
+    arr := [1, 'two', 3];
+end.