+++ examples/test/dynarray/test_dynarray_literal_empty.pas
@@ -1,0 +1,8 @@
+program TestDynArrayLiteralEmpty;
+var
+    arr: array of integer;
+begin
+    arr := [];
+    writeln(Length(arr));  { 0 }
+    writeln(arr = nil);    { TRUE }
+end.