+++ src/pascalc/ast_printer.c
@@ -440,6 +440,17 @@ void print_ast(ASTNode *node, int indent) {
             break;
         }
 
+        case NODE_DYNARRAY_LITERAL: {
+            printf("[Dynamic Array Literal] (elem type=%d)\n", node->data.var_idx);
+            int elem_num = 1;
+            for (ASTNode *elem = node->left; elem; elem = elem->next) {
+                print_indent(indent + 1);
+                printf("Elem %d:\n", elem_num++);
+                print_ast(elem, indent + 2);
+            }
+            break;
+        }
+
         case NODE_SET_IN:
             printf("[Set In]\n");
             print_indent(indent + 1);